requirements.txt
Django==1.11.3
mistune==0.7.4
django-solo==1.1.2
django-taggit==0.22.0
factory-boy==2.8.1
django-flat-responsive==1.4.1

# up requirements
pyyaml
requests
dj_database_url
